+++ components/molecules/UserBar.tsx
@@ -6,7 +6,7 @@ interface User {
   updatedAt: string
 }
 
-export const UserBar: React.VFC<User> = ({ username, updatedAt }) => {
+const UserBar: React.VFC<User> = ({ username, updatedAt }) => {
   return (
     <div className="w-full flex items-center justify-start bg-white p-2 space-x-2 rounded-md">
       <UserIcon />
@@ -20,3 +20,5 @@ export const UserBar: React.VFC<User> = ({ username, updatedAt }) => {
     </div>
   )
 }
+
+export default UserBar
